Folder structure reorganized for v1.0.0
